requirements.txt
# Core Python stack
numpy
pandas
scipy

# Deep Learning
torch
torchvision
torchaudio

# NLP (BERT)
transformers
tokenizers

# Graph Neural Networks
torch-geometric

# Machine Learning & Evaluation
scikit-learn
optuna

# Image Processing
Pillow
opencv-python

# Data handling & utilities
tqdm
requests

# Visualization & analysis
matplotlib
seaborn
